+++ .gitignore
@@ -32,6 +32,7 @@ build/
 .DS_Store
 
 # Python
+dist/
 poetry.lock
 __pycache__
