+++ packages/sshnoports_python/pyproject.toml
@@ -1,9 +1,12 @@
 [tool.poetry]
 name = "sshnpdpy"
-version = "0.1.0"
+version = "0.3.0"
 description = ""
-authors = [""]
+authors = ["Xavier Lin <@Xlin123>"]
 readme = "README.md"
+
+[tool.poetry.scripts]
+sshnpdpy = 'sshnpdpy:main'
 
 [tool.poetry.dependencies]
 python = "^3.10"
@@ -14,6 +17,8 @@ paramiko = "3.3.1"
 pycparser = "^2.21"
 PyNaCl = "^1.5.0"
 setuptools = "^69.0.0"
+
+
 
 
 [build-system]
