--- packages/sshnpdpy/lib/__init__.py
@@ -1,1 +1,0 @@
-from .sshnpdclient import SSHNPDClient
